Update to fumadocs 10.0.2
