+++ next.config.mjs
@@ -1,20 +1,18 @@
 import createMDX from "fumadocs-mdx/config";
-import nextBuildId from "next-build-id";
 
 const withMDX = createMDX({
   rootMapPath: "./src/_map.ts",
   rootContentPath: "./content",
+  buildSearchIndex: {
+    filter: (path) => {
+      return path.startsWith("docs");
+    },
+  },
 });
 
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   pageExtensions: ["ts", "tsx", "mdx"],
-  generateBuildId: () => {
-    const id = nextBuildId.sync({ dir: process.cwd() });
-
-    process.env.NEXT_BUILD_ID = id;
-    return id;
-  },
   images: {
     unoptimized: true,
   },
