+++ package.json
@@ -4,7 +4,7 @@
   "private": true,
   "scripts": {
     "dev": "next dev",
-    "build": "next build && next-sitemap --config next-sitemap.config.js",
+    "build": "next build && next-sitemap --config next-sitemap.config.js && node ./scripts/add-to-assets.mjs",
     "start": "next start -H 0.0.0.0",
     "lint": "next lint"
   },
@@ -12,11 +12,11 @@
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.0",
     "framer-motion": "^11.0.8",
-    "fumadocs-core": "^10.0.0",
-    "fumadocs-mdx": "^8.1.1",
-    "fumadocs-ui": "^10.0.0",
+    "fumadocs-core": "^10.0.2",
+    "fumadocs-mdx": "^8.2.1",
+    "fumadocs-ui": "^10.0.2",
     "lucide-react": "^0.294.0",
-    "next": "^14.1.1",
+    "next": "^14.1.3",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-twitter-embed": "^4.0.4",
@@ -26,28 +26,22 @@
   "devDependencies": {
     "@types/eslint": "^8.56.5",
     "@types/mdx": "^2.0.11",
-    "@types/node": "^20.11.24",
-    "@types/react": "^18.2.61",
-    "@types/react-dom": "^18.2.19",
+    "@types/node": "^20.11.25",
+    "@types/react": "^18.2.64",
+    "@types/react-dom": "^18.2.21",
     "@typescript-eslint/eslint-plugin": "^6.21.0",
     "@typescript-eslint/parser": "^6.21.0",
     "autoprefixer": "^10.4.18",
     "eslint": "8.55.0",
-    "eslint-config-next": "^14.1.1",
+    "eslint-config-next": "^14.1.3",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-react": "7.33.2",
-    "next-build-id": "^3.0.0",
     "next-sitemap": "^4.2.3",
     "postcss": "^8.4.35",
     "prettier": "^3.2.5",
-    "prettier-plugin-tailwindcss": "^0.5.11",
+    "prettier-plugin-tailwindcss": "^0.5.12",
     "sharp": "^0.33.2",
     "tailwindcss": "^3.4.1",
     "typescript": "5.3.2"
-  },
-  "pnpm": {
-    "patchedDependencies": {
-      "fumadocs-core@10.0.0": "patches/fumadocs-core@10.0.0.patch"
-    }
   }
 }
